requirements.txt
aioesphomeapi==18.5.2
ajsonrpc==1.2.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.2.0
argcomplete==3.2.1
asyncio==3.4.3
bitarray==2.9.0
bitstring==4.1.4
blinker==1.6.2
bottle==0.12.25
certifi==2023.11.17
cffi==1.16.0
chacha20poly1305-reuseable==0.12.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
ecdsa==0.18.0
esphome==2023.11.6
esphome-dashboard==20231107.0
esptool==4.6.2
Flask==2.3.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kconfiglib==13.7.1
Mako==1.3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
mysql-connector-python==8.1.0
noiseprotocol==0.3.1
numpy==2.1.1
openai==1.37.1
packaging==24.1
paho-mqtt==1.6.1
platformio==6.1.11
protobuf==4.21.12
psycopg2-binary==2.9.9
pycairo==1.27.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyelftools==0.29
PyGObject==3.50.0
PyJWT==2.8.0
pyparsing==3.1.1
pyserial==3.5
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
reedsolo==1.7.0
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.31.1
tabulate==0.9.0
tornado==6.3.3
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.23.2
voluptuous==0.13.1
Werkzeug==2.3.0
wsproto==1.2.0
zeroconf==0.123.0
